Tic Tac Toe
This is the Tic Tac Toe which is Game based application. In this game there are 2 players, both have to play turn by turn , A functionality is provided to them , X and O . So, the one who make 3 same signs in same row or same column or in same digonal wins the Game. If noone does so, then the game finally Draws...